% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{addMethCol}
\alias{addMethCol}
\title{Adding methyl column that has proportion of reads that were methylated for 
each site.
Note: Assigns methyl column by reference with ":="}
\usage{
addMethCol(BSDTList)
}
\arguments{
\item{BSDTList}{A bisulfite datatable or list of datatables with a column for
number of methylated reads (hitCount) and a column for number of total reads 
(readCount) for each cytosine that was measured.}
}
\value{
The BSDTList but with extra methyl column on each data.table in list.
}
\description{
Adding methyl column that has proportion of reads that were methylated for 
each site.
Note: Assigns methyl column by reference with ":="
}
\examples{
data("GM06990_1_ExampleSet",package="MIRA") #exampleBSDT
exampleBSDT[,methyl := NULL] #removing methyl column
addMethCol(list(exampleBSDT))
}
