% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annoDF}
\alias{annoDF}
\title{A data.table with annotation for a bisulfite sample (name and sample type)}
\format{A data.table, data.frame object}
\usage{
data(annoDF)
}
\value{
No return value.
}
\description{
A data.table with annotation for a bisulfite sample (name and sample type)
}
\keyword{datasets}
