% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigBinDT}
\alias{bigBinDT}
\title{A data.table with binned methylation data for the "Applying MIRA to a 
Biological Question" vignette. Has data for 24 Ewing samples and 24
muscle-related samples (human skeletal myoblasts and other related
samples).}
\format{A data.table, data.frame object}
\usage{
data(ewingMyobigBinDT)
}
\value{
No return value.
}
\description{
A data.table with binned methylation data for the "Applying MIRA to a 
Biological Question" vignette. Has data for 24 Ewing samples and 24
muscle-related samples (human skeletal myoblasts and other related
samples).
}
\keyword{datasets}
