% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigBinDT1}
\alias{bigBinDT1}
\title{A data.table with binned methylation data for the first part of
the "Applying MIRA to a Biological Question" vignette. 
Has data for 3 Ewing samples and 3 muscle-related 
samples (human skeletal myoblasts and other related
samples). It is the product of `aggregateMethyl`.}
\format{A data.table, data.frame object}
\usage{
data(ewingMyobigBinDT1)
}
\value{
No return value.
}
\description{
A data.table with binned methylation data for the first part of
the "Applying MIRA to a Biological Question" vignette. 
Has data for 3 Ewing samples and 3 muscle-related 
samples (human skeletal myoblasts and other related
samples). It is the product of `aggregateMethyl`.
}
\keyword{datasets}
