% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{binProcess}
\alias{binProcess}
\title{Convenience function to aggregate bisulfite scores across bins, across regions.
You need a list of ranges (a bed file as a data.table, in a list)
and a list of BSDTs.}
\usage{
binProcess(rangeDTList, BSDTNames, anno, binCount = 11)
}
\arguments{
\item{rangeDTList}{This is a list of data.tables, with each just having chr,
start and end.}

\item{anno}{An annotation table (data.table), keyed with the same key as the
BSDTs; also with a `cell_type` column, used to append cell_type to the binned
result tables.}
}
\description{
Convenience function to aggregate bisulfite scores across bins, across regions.
You need a list of ranges (a bed file as a data.table, in a list)
and a list of BSDTs.
}
