% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{BSBinAggregate}
\alias{BSBinAggregate}
\title{A wrapper of BSAggregate that first bins regions and then aggregates
each bin across a set of regions, individually.}
\usage{
BSBinAggregate(BSDT, rangeDT, binCount, minReads = 500,
  byRegionGroup = TRUE, splitFactor = "id")
}
\arguments{
\item{rangeDT}{A data table with the sets of regions to be binned, 
with columns named start, end}

\item{binCount}{Number of bins across the region}

\item{minReads}{Filter out bins with fewer than X reads before returning.}

\item{byRegionGroup}{Pass along to binCount (see ?binCount)}
}
\description{
Produced originally for binning Ewing RRBS data across various region sets
}
