% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{BSBinPlots}
\alias{BSBinPlots}
\title{Given a binned BSDT (which is produced from BSBinAggregate), this plot will
produce and return a few ggplots (which could then be printed).
It produces a combined, and a faceted by cell-type plot, showing how signal
varies across bins.}
\usage{
BSBinPlots(binnedBSDT, binCount, regionName = "regions")
}
\description{
TODO: It's EWS-centric; needs to become parameterized.
}
