% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{BSdtToGRanges}
\alias{BSdtToGRanges}
\title{Converts a list of data.tables into GRanges.}
\usage{
BSdtToGRanges(dtList)
}
\arguments{
\item{dtList}{A list of data.tables,
Each must have "chr","start","hitCount", and "readCount" columns}
}
\value{
a list of GRanges objects, strand has been set to "*",
"start" and "end" have both been set to "start" of the DT
}
\description{
Converts a list of data.tables into GRanges.
}
