% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{BSFilter}
\alias{BSFilter}
\title{Given a BSDT (bisulfite data.table), remove any entries that overlap
regions given in the excludeGR argument and/or filter out sites
that have lower than a minimum number of reads.}
\usage{
BSFilter(BSDT, minReads = 10, excludeGR = NULL)
}
\arguments{
\item{BSDT}{Bisulfite data.table to filter}

\item{minReads}{Require at least this level of coverage at a cpg.}

\item{excludeGR}{GRanges object with regions to filter.}
}
\value{
The BSDT with appropriate regions removed.
}
\description{
Given a BSDT (bisulfite data.table), remove any entries that overlap
regions given in the excludeGR argument and/or filter out sites
that have lower than a minimum number of reads.
}
