% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{BSmultiScore}
\alias{BSmultiScore}
\title{Wrapper of BSBinAggregate to do multiple regions (ie multiple TFs) and/or multiple samples}
\usage{
BSmultiScore(BSDTList, rangeDTList, binNum = 11)
}
\arguments{
\item{BSDTList}{This should be a list of Bisulfite data.tables that contains a methyl column with
chr, start, hitCount (number of reads methylated), readCount (number of reads total), and 
methyl (percent methylation present) columns.}

\item{rangeDTList}{This is a list of genomic ranges, each corresponding to a different feature; 
it should be a list of data.tables but a GRangesList will be accepted.}

\item{binNumber}{gives number of bins to divide each region into.}
}
\description{
Wrapper of BSBinAggregate to do multiple regions (ie multiple TFs) and/or multiple samples
}

