% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realign.R
\name{build_bt_index}
\alias{build_bt_index}
\title{Build the bowtie index}
\usage{
build_bt_index(cDNAFile, wdir, pipedir, iDir, prefix = "index",
  verbose = FALSE)
}
\arguments{
\item{cDNAFile}{file with cDNA sequences in fasta format}

\item{wdir}{working directory}

\item{pipedir}{pipeline directory}

\item{iDir}{directory for writing the bowtie index}

\item{prefix}{prefix string for index file}
}
\value{
path to bowtie index
}
\description{
Build the bowtie index
}
