% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{buildJ}
\alias{buildJ}
\title{helper function
given a vector of colums, and the equally-sized vector of functions
to apply to those columns, constructs a j-expression for use in
a data.table.
use it in a DT[,eval(parse(text=buildJ(cols,funcs)))]}
\usage{
buildJ(cols, funcs)
}
\description{
helper function
given a vector of colums, and the equally-sized vector of functions
to apply to those columns, constructs a j-expression for use in
a data.table.
use it in a DT[,eval(parse(text=buildJ(cols,funcs)))]
}
