% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpression.R
\name{diffExpression}
\alias{diffExpression}
\title{Differential Expression}
\usage{
diffExpression(x, eset, annotation, wdir, volcanoplots = F)
}
\arguments{
\item{volcanoplots}{}
}
\description{
Differential Expression
}
