% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadDataSet}
\alias{downloadDataSet}
\title{Download a microarray dataset from ArrayExpress or GEO}
\usage{
downloadDataSet(x, wdir)
}
\arguments{
\item{wdir}{}
}
\description{
Download a microarray dataset from ArrayExpress or GEO
}
