% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadPlatformFile}
\alias{downloadPlatformFile}
\title{Download a microarray platform file}
\usage{
downloadPlatformFile(eid, wdir)
}
\arguments{
\item{wdir}{}
}
\value{
full path filename of the platform file
}
\description{
Download a microarray platform file
}
