% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{dtToGr}
\alias{dtToGr}
\title{Convert a data.table to GRanges object.}
\usage{
dtToGr(DT, chr = "chr", start = "start", end = NA, strand = NA,
  name = NA, splitFactor = NA, metaCols = NA)
}
\arguments{
\item{DT}{a data.table with at least "chr" and "start" columns}
}
\value{
gr A genomic ranges object derived from DT
}
\description{
Convert a data.table to GRanges object.
}
