% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{dtToGrInternal}
\alias{dtToGrInternal}
\title{Internal part of a utility to convert data.tables into GRanges objects
genes = dtToGR(gModels, "chr", "txStart", "txEnd", "strand", "geneId").}
\usage{
dtToGrInternal(DT, chr, start, end = NA, strand = NA, name = NA,
  metaCols = NA)
}
\arguments{
\item{DT}{A data.table with at least "chr" and "start" columns}
}
\value{
gr A genomic ranges object derived from DT
}
\description{
Internal part of a utility to convert data.tables into GRanges objects
genes = dtToGR(gModels, "chr", "txStart", "txEnd", "strand", "geneId").
}
