% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleAnnoDT1}
\alias{exampleAnnoDT1}
\title{A data.table with annotation for a bisulfite sample.}
\format{A data.table, data.frame object}
\usage{
data(exampleAnnoDT1)
}
\description{
Has columns for name and sample type.
}
\keyword{datasets}
