% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleAnnoDT2}
\alias{exampleAnnoDT2}
\title{Annotation data for the "Applying MIRA to a 
Biological Question" vignette.}
\format{A data.table, data.frame object}
\usage{
data(exampleAnnoDT2)
}
\description{
A data.table with annotation data (sampleName
and sampleType columns) for 24 Ewing samples and 24 muscle-related
samples (human skeletal myoblasts and other related samples).
}
\keyword{datasets}
