% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleBins}
\alias{exampleBins}
\title{A data.table with artificial binned methylation data}
\format{A data.table, data.frame object}
\usage{
data(exampleBins)
}
\value{
No return value.
}
\description{
A data.table with artificial binned methylation data
}
\keyword{datasets}
