% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleBSDT}
\alias{exampleBSDT}
\title{A data.table containing DNA methylation data.}
\format{A data.table, data.frame object}
\usage{
data(exampleBSDT)
}
\description{
This object contains reduced
representation bisulfite sequencing data from the
GM06990 B-lymphoblastoid cell line that was generated by the ENCODE
project (GEO accession GSE27584, sample GSM980583). Has columns for
position (chr, start), number methylated reads, number of reads total,
and sample name.
Only a subset of the full data is used in this object and the 
data was switched from hg19 to hg38.
}
\keyword{datasets}
