% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleBSseqObj}
\alias{exampleBSseqObj}
\title{A BSseq object which has the top 20 rows of the BSseq object "BS.chr22"
which was included as built-in data in the bsseq package}
\format{A BSseq object}
\usage{
data(exampleBSseqObj)
}
\description{
A BSseq object which has the top 20 rows of the BSseq object "BS.chr22"
which was included as built-in data in the bsseq package
}
\keyword{datasets}
