% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleRegionSet}
\alias{exampleRegionSet}
\title{A GRanges object with a subset of ranges corresponding to Nrf1 binding}
\format{A GRanges object}
\usage{
data(exampleRegionSet)
}
\value{
No return value.
}
\description{
A GRanges object with a subset of ranges corresponding to Nrf1 binding
}
\keyword{datasets}
