% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{extractSampleName}
\alias{extractSampleName}
\title{extract sample names from file names as the first part of the file name (before any suffix)}
\usage{
extractSampleName(fileNames, suffixSep = "\\\\.", pathSep = "/")
}
\description{
extract sample names from file names as the first part of the file name (before any suffix)
}

