% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{goEnrichment}
\alias{goEnrichment}
\title{Run GO enrichment}
\usage{
goEnrichment(dt, transcript2GO, nTopNodes, verbose)
}
\arguments{
\item{verbose}{}
}
\value{
list of GO enrichments, one element per contrast
}
\description{
Run GO enrichment
}
