% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{goEnrichmentHelper}
\alias{goEnrichmentHelper}
\title{Calculate GO enrichment for one contrast}
\usage{
goEnrichmentHelper(dtCol, transcript2GO, nTopNodes = 20)
}
\arguments{
\item{nTopNodes}{}
}
\value{
data frame of enrichment results for one contrast
}
\description{
Calculate GO enrichment for one contrast
}
