% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{grToDt}
\alias{grToDt}
\title{convert a GenomicRanges into a data.table}
\usage{
grToDt(GR)
}
\arguments{
\item{GR}{A GRanges object}
}
\value{
A data.table object.
}
\description{
convert a GenomicRanges into a data.table
}
