% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{loadAnnotation}
\alias{loadAnnotation}
\title{Loads GO annotation from Biomart}
\usage{
loadAnnotation(specimen, annoFile = NULL)
}
\arguments{
\item{annoFile}{}
}
\value{
list of gene id to go mappings:
listnames are genes
list values are vectors of GO IDs
}
\description{
Loads GO annotation from Biomart
}
