% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realign.R
\name{mapProbes}
\alias{mapProbes}
\title{map probe sequences to reference genome}
\usage{
mapProbes(x, wdir, pipedir, prefix = "index", nCores = 4)
}
\arguments{
\item{x}{named vector containing parameters
cDNAFile
platformFile
probeIdColName
probeSeqColName}

\item{wdir}{STRING working directory}

\item{pipedir}{STRING pipeline directory}

\item{prefix}{STRING}

\item{nCores}{INT number of cores bowtie should use}
}
\description{
map probe sequences to reference genome
}
