% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA-package.R
\docType{package}
\name{MIRA-package}
\alias{MIRA}
\alias{MIRA-package}
\title{MIRA: Bayesian Multilevel Inference for Longitudinal Data}
\description{
MIRA provides data preparation, configurable priors, Stan fitting, and
posterior summaries for continuous longitudinal outcomes. Frequentist
exploration and reproducible report bundles are also available.
}
\details{
Start with \code{\link[=mira_data_long]{mira_data_long()}}, configure priors with \code{\link[=mira_prior_long]{mira_prior_long()}},
fit with \code{\link[=mira_fit_long]{mira_fit_long()}}, and inspect \code{\link[=mira_summary_long]{mira_summary_long()}}. Bayesian
fitting requires CmdStan and a C++ toolchain. For data exploration use
\code{\link[=mira_info_long]{mira_info_long()}} and \code{\link[=mira_report_freq_long]{mira_report_freq_long()}}.
}
\seealso{
\url{https://github.com/davide-celani/MIRA}
}
\author{
\strong{Maintainer}: Davide Celani \email{davide.celani.00@gmail.com}

Authors:
\itemize{
  \item Davide Celani \email{davide.celani.00@gmail.com}
}

}
\keyword{internal}
