% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\docType{package}
\name{MIRA}
\alias{MIRA}
\alias{MIRA-package}
\title{Methylation-based Inference of Regulatory Activity (MIRA)}
\description{
MIRA is a score that infers regulatory activity of genomic elements
based on DNA methylation data. It assess the degree of dip in methylation
level surrounding a regulatory site of interest, such as 
transcription factor binding sites.
This package provides functions for aggregating methylation 
data across region sets, in bins.
}
\references{
\url{http://github.com/databio}
}
\author{
Nathan Sheffield

John Lawson
}
