% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mira_prior_long.R
\name{mira_prior_bcva_long}
\alias{mira_prior_bcva_long}
\title{Create BCVA priors}
\usage{
mira_prior_bcva_long(
  stan_data,
  informativeness = "standard",
  time_unit = "months",
  ...
)
}
\arguments{
\item{stan_data}{Data prepared from BCVA columns.}

\item{informativeness}{\code{"standard"}, \code{"weak"}, \code{"informative"}, or
\code{"custom"}.}

\item{time_unit}{Unit used by \code{stan_data$time_value}.}

\item{...}{Optional prior replacements passed to \code{\link[=mira_prior_long]{mira_prior_long()}}.}
}
\value{
A \code{mira_prior_long} object.
}
\description{
Create BCVA priors
}
