% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mira_prior_long.R
\name{mira_prior_stan_data_long}
\alias{mira_prior_stan_data_long}
\title{Convert MIRA priors to Stan data}
\usage{
mira_prior_stan_data_long(prior)
}
\arguments{
\item{prior}{A \code{mira_prior_long} object or complete named list.}
}
\value{
A list containing exactly the 16 prior fields declared in Stan.
}
\description{
Convert MIRA priors to Stan data
}
