% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{mira_stan_file}
\alias{mira_stan_file}
\title{Get the Stan file associated with a MIRA model}
\usage{
mira_stan_file(model = "gaussian_longitudinal")
}
\arguments{
\item{model}{Character string identifying the MIRA model.}
}
\value{
Path to the Stan file.
}
\description{
Get the Stan file associated with a MIRA model
}
\keyword{internal}
