% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mira_prior_long.R
\name{mira_validate_prior_long}
\alias{mira_validate_prior_long}
\title{Validate a MIRA prior specification}
\usage{
mira_validate_prior_long(prior)
}
\arguments{
\item{prior}{A \code{mira_prior_long} object or complete named list.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Validate a MIRA prior specification
}
