% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{normalizeEwingsToNonEwings}
\alias{normalizeEwingsToNonEwings}
\title{Calculate the median methylation in each bin, for all non-EWS samples, and
then divide methylation by this median value to get a relative increase/decrease
in each bin.}
\usage{
normalizeEwingsToNonEwings(binnedBSDT)
}
\description{
Calculate the median methylation in each bin, for all non-EWS samples, and
then divide methylation by this median value to get a relative increase/decrease
in each bin.
}
