% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normalizeLimma}
\alias{normalizeLimma}
\title{Title}
\usage{
normalizeLimma(filenames, fType, dualChannel, bgmethod = "normexp",
  treatDualChannelAsSingleChannel = F, verbose = F)
}
\arguments{
\item{treatDualChannelAsSingleChannel}{}
}
\description{
Title
}
