% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{normalizeMIRA}
\alias{normalizeMIRA}
\title{Function to normalize case/experimental samples to the controls}
\usage{
normalizeMIRA(binnedDT)
}
\arguments{
\item{binnedDT}{A datatable containing bins for each region set for each sample;bins contain
aggregated methylation across regions for that sample; 
it should have a column with sample annotation so cases and controls can be split up as
well as annotation of the region sets (featureID column) and sampleType column (case/control).}
}
\value{
binnedDT The input DT but with normalized "methyl" values
}
\description{
It finds the median of the controls for each bin for each region set
then divides by the median (oldBinVal/medianBinVal) for each bin.
}
