% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{parseBiseq}
\alias{parseBiseq}
\title{Takes a data.table from BSreadBiSeq and parses the strange x/y format
of methylation calls, splitting them into individual columns}
\usage{
parseBiseq(DT)
}
\arguments{
\item{DT}{data.table to parse
"chr", "start", "end", "meth", "rate", "strand" columns expected
in that order.}
}
\value{
data.table with separate methylated and unmethylated columns.
Specific col names are set
}
\description{
Takes a data.table from BSreadBiSeq and parses the strange x/y format
of methylation calls, splitting them into individual columns
}
