% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mira_prior.R
\name{print.mira_prior}
\alias{print.mira_prior}
\title{Print a MIRA prior specification}
\usage{
\method{print}{mira_prior}(x, ...)
}
\arguments{
\item{x}{A \code{mira_prior} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print a MIRA prior specification
}
