% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mira_prior_long.R
\name{print.mira_prior_long}
\alias{print.mira_prior_long}
\title{Print a MIRA prior specification}
\usage{
\method{print}{mira_prior_long}(x, ...)
}
\arguments{
\item{x}{A \code{mira_prior_long} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print a MIRA prior specification
}
