% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{returnMIRABins}
\alias{returnMIRABins}
\title{Function to aggregate methylation data into bins over all regions in each region set;}
\usage{
returnMIRABins(BSDT, GRList, binNum = 11, minReads = 500,
  sampleNameInBSDT = TRUE, sampleType = NULL)
}
\arguments{
\item{BSDT}{A single data table that has DNA methylation data including :.}

\item{GRList}{A GRangesList object containing region sets, each set corresponding to a regulatory element;
Each regionSet in the list should be named; A named list of data.tables also works.}

\item{binNum}{How many bins each region should be split into for aggregation of the DNA methylation data}

\item{minReads}{Filter out bins with fewer than X reads before returning.}

\item{sampleNameInBSDT}{boolean for whether the BSDT has a sampleName column}

\item{sampleType}{could be case/control, tissue type, etc.}
}
\value{
a data.table with binNum rows for each region set containing aggregated methylation data
Each region was split into bins; methylation was put in these bins; 
Output contains sum of the all corresponding bins for the regions of each region set 
ie for all regions in each region set: first bins summed, second bins summed, etc
}
\description{
Function to aggregate methylation data into bins over all regions in each region set;
}
\examples{
#loading necessary libraries
library(data.table)
library(ggplot2)
library(GenomicRanges)
library(MIRA)

#setting initial parameter
binNum=11

#reading in annotation
annoDF=read.csv("C:/cygwin64/Comp_Epigenetics/Bcell_epithelialAnno.csv")
annoDF=as.data.table(annoDF)

#reading in region data
#reading in bisulfite data
#will generalize/replace these later
load("C:/cygwin64/Comp_Epigenetics/dump/BSDTListBcellEpithelial.RData")
load("C:/cygwin64/Comp_Epigenetics/dump/encodeRegionSampling.RData")
BSDTList=BSDTList[c(1:3,11:13)]
#save(BSDTList,file="C:/cygwin64/Comp_Epigenetics/MIRA_Projects/BSDTListBcellepithelial3_3.RData")
someFeatures=someFeatures[1:6]

#converting to proper format
BSDTList=addMethCol(BSDTList = BSDTList)


#doing MIRA analysis
bigBin=lapply(X = BSDTList,FUN = returnMIRABins,GRList=someFeatures,binNum=11,sampleNameInBSDT=TRUE)

bigBinDT=rbindlist(bigBin)#need to make sure that sample names have a column in here

#adding sampleType from annotation object
setkey(bigBinDT,sampleName)
setkey(annoDF,sampleName)
bigBinDT=merge(bigBinDT,annoDF, all.x=TRUE)

#scoring samples
sampleScores=bigBinDT[,.(score = scoreDip(methyl,binNum)),by=.(featureID,sampleName)]
setkey(sampleScores,sampleName)
setkey(annoDF,sampleName)
sampleScores=merge(sampleScores,annoDF,all.x=TRUE)



#visualizing MIRA signatures
plotMIRARegions(binnedRegDT = bigBinDT)

#visualizing MIRA scores
plotMIRAScores(scoreDT=sampleScores)


}
