% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanFile.R
\name{scanFileType}
\alias{scanFileType}
\title{Reads the magic number from a binary file and identifies the microarray file type}
\usage{
scanFileType(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Reads the magic number from a binary file and identifies the microarray file type
}
