% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeLoci}
\alias{summarizeLoci}
\title{Summarise intensity values for one gene locus}
\usage{
summarizeLoci(eid, alFile, normEset, wdir, sumFUN)
}
\arguments{
\item{eid}{"experiment identifier"}

\item{alFile}{"alignment file"}

\item{normEset}{"normalised Expressionset"}

\item{wdir}{"working directory"}

\item{sumFUN}{"summarsation function, usually median or mean"}
}
\description{
Summarise intensity values for one gene locus
}
