% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{writeEnrichment}
\alias{writeEnrichment}
\title{Write a single enrichment table to a file}
\usage{
writeEnrichment(idx, resList, wdir)
}
\arguments{
\item{wdir}{}
}
\description{
Write a single enrichment table to a file
}
