% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realign.R
\name{writeProbesAsFasta}
\alias{writeProbesAsFasta}
\title{Write Probe sequences into fasta file for all microarray definition files using tabular text format}
\usage{
writeProbesAsFasta(x, wdir)
}
\arguments{
\item{wdir}{}
}
\description{
Write Probe sequences into fasta file for all microarray definition files using tabular text format
}
