% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realign.R
\name{writeTagFasta}
\alias{writeTagFasta}
\title{Write probe sequences into a fasta file for celera genomics microarray files}
\usage{
writeTagFasta(seqs, outFile)
}
\arguments{
\item{outFile}{}
}
\description{
Write probe sequences into a fasta file for celera genomics microarray files
}
