% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{addMethCol}
\alias{addMethCol}
\title{Adding methyl column that has proportion of reads that were methylated for each site.
Note: Assigns methyl column by reference with ":="}
\usage{
addMethCol(BSDTList)
}
\arguments{
\item{BSDTList}{A bisulfite datatable or list of datatables with a column for number of methylated 
reads (hitCount) and a column for number of total reads (readCount) for each cytosine that 
was measured.}
}
\description{
Adding methyl column that has proportion of reads that were methylated for each site.
Note: Assigns methyl column by reference with ":="
}
