% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{aggregateMethyl}
\alias{aggregateMethyl}
\title{Function to aggregate methylation data into bins 
over all regions in each region set.}
\usage{
aggregateMethyl(BSDT, GRList, binNum = 11, minReads = 500,
  sampleNameInBSDT = TRUE)
}
\arguments{
\item{BSDT}{A single data.table that has DNA methylation data on individual 
sites. With columns: "chr" for chromosome, "start" for 
cytosine coordinate, "methylProp" for proportion of 
methylation (0 to 1), "methylCount" for number of methylated reads, and
"coverage" for total number of reads.
In addition, a "sampleName" column is preferred.}

\item{GRList}{A GRangesList object containing region sets, each set 
corresponding to a type of regulatory element. Each regionSet in the list should
be named. A named list of data.tables also works.}

\item{binNum}{How many bins each region should be split into for aggregation 
of the DNA methylation data.}

\item{minReads}{Filter out bins with fewer than minReads reads.}

\item{sampleNameInBSDT}{boolean for whether the BSDT has a sampleName column}
}
\value{
a data.table with binNum rows for each region set containing
aggregated methylation data.
Each region was split into bins; methylation was put in these bins; 
Output contains sum of the all corresponding bins for the regions of each 
region set, ie for all regions in each region set: first bins summed, second 
bins summed, etc. Columns of the output should be "regionGroupID", "methylProp", 
"coverage", "featureID", and possibly "sampleName".
For information on symmetry of bins and output when a region set has
strand info, see ?BSBinAggregate.
}
\description{
Function to aggregate methylation data into bins 
over all regions in each region set.
}
\examples{
data("exampleBSDT", package = "MIRA")
data("exampleRegionSet", package = "MIRA")
exBinDT = aggregateMethyl(exampleBSDT, exampleRegionSet)
}
