% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{BSreadBiSeq}
\alias{BSreadBiSeq}
\title{Read in files from biseq meth caller}
\usage{
BSreadBiSeq(files, contrastList = NULL,
  sampleNames = extractSampleName(files), cores = 4, returnAsList = FALSE)
}
\arguments{
\item{files}{a list of filenames (use parseInputArg if necessary)}

\item{contrastList}{a list of named character vectors, each with length equal to the number of items in files. These will translate into column names in the final table.}

\item{sampleNames}{a vector of length length(files), name for each file. You can also just use contrastList to implement the same thing so this is really unnecessary...}

\item{cores}{number of processors.}

\item{returnAsList}{Whether to return the output as a list or as one big data.table}
}
\value{
Data from each input file joined together into one big data.table.
If returnAsList=TRUE, then input from each file will be in its own data.table in a list.
}
\description{
Read in files from biseq meth caller
}
