% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleAnnoDT1}
\alias{exampleAnnoDT1}
\title{A data.table with annotation for a bisulfite sample (name and sample type)}
\format{A data.table, data.frame object}
\usage{
data(exampleAnnoDT1)
}
\description{
A data.table with annotation for a bisulfite sample (name and sample type)
}
\keyword{datasets}
