% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleBSDT}
\alias{exampleBSDT}
\title{A data.table containing methylation data 
(position, number methylated reads and number of reads total)
Also includes name of the sample it was derived from.}
\format{A data.table, data.frame object}
\usage{
data(GM06990_1_ExampleSet)
}
\value{
No return value.
}
\description{
A data.table containing methylation data 
(position, number methylated reads and number of reads total)
Also includes name of the sample it was derived from.
}
\keyword{datasets}
