% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{grToDt}
\alias{grToDt}
\title{convert a GenomicRanges into a data.table}
\usage{
grToDt(GR, includeStrand = FALSE)
}
\arguments{
\item{GR}{A GRanges object}

\item{includeStrand}{Boolean, whether to include strand from GR in output DT}
}
\value{
A data.table object with columns "chr","start", and "end" (possibly strand)
}
\description{
convert a GenomicRanges into a data.table
}
