% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\docType{package}
\name{MIRA}
\alias{MIRA}
\alias{MIRA-package}
\title{Methylation-based Inference of Regulatory Activity (MIRA)
MIRA is a score that measures the degree of dip in methylation
level surrounding a regulatory site of interest, such as a 
transcription factor binding sites.
This script provides functions for aggregating methylation 
data across region sets, in bins.}
\description{
Methylation-based Inference of Regulatory Activity (MIRA)
MIRA is a score that measures the degree of dip in methylation
level surrounding a regulatory site of interest, such as a 
transcription factor binding sites.
This script provides functions for aggregating methylation 
data across region sets, in bins.
}
\references{
\url{http://github.com/databio}
}
\author{
Nathan Sheffield
}
