% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mira_fit}
\alias{mira_fit}
\title{Fit MIRA longitudinal model}
\usage{
mira_fit(
  stan_data,
  prior = mira_prior(stan_data, profile = "default"),
  chains = 4,
  parallel_chains = chains,
  iter_warmup = 1000,
  iter_sampling = 3000,
  seed = 123,
  refresh = 100
)
}
\arguments{
\item{stan_data}{Data prepared for Stan.}

\item{prior}{A \code{mira_prior} object. If omitted, the default
MIRA prior specification is used.}

\item{chains}{Number of MCMC chains.}

\item{parallel_chains}{Number of parallel chains.}

\item{iter_warmup}{Number of warmup iterations.}

\item{iter_sampling}{Number of sampling iterations.}

\item{seed}{Random seed.}

\item{refresh}{Number of iterations between progress messages.}
}
\value{
A CmdStanMCMC object.
}
\description{
Fits the longitudinal Student-t model using CmdStan.
}
