% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mira_prior.R
\name{mira_prior_stan_data}
\alias{mira_prior_stan_data}
\title{Convert MIRA priors to Stan data}
\usage{
mira_prior_stan_data(prior)
}
\arguments{
\item{prior}{A \code{mira_prior} object.}
}
\value{
A named list suitable for CmdStan.
}
\description{
Convert MIRA priors to Stan data
}
