% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mira_prior.R
\name{mira_validate_prior}
\alias{mira_validate_prior}
\title{Validate a MIRA prior specification}
\usage{
mira_validate_prior(prior)
}
\arguments{
\item{prior}{A \code{mira_prior} object.}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate a MIRA prior specification
}
