% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{MIRAScore}
\alias{MIRAScore}
\title{Get MIRA score for each sample/region set combination}
\usage{
MIRAScore(BSDT, GRList, binNum = 11, scoringMethod = "logRatio",
  minReads = 500)
}
\arguments{
\item{BSDT}{A single data table that has DNA methylation data on individual 
sites including a "chr" column with chromosome, a "start" column with the 
coordinate number for the cytosine, a "methylProp" column with proportion of 
methylation (0 to 1), optionally a "methylCount" column with 
number of methylated reads 
for each site, optionally a "coverage" column with 
total number of reads for each 
site, and a "sampleName" column with a sample identifier/name (required).}

\item{GRList}{A GRangesList object containing region sets, each set 
corresponding to a regulatory element (or having regions with the 
same biological annotation).
Each regionSet in the list should be named.}

\item{binNum}{How many bins each region should be split into for aggregation 
of the DNA methylation data.}

\item{scoringMethod}{Method to calculate MIRA score after binning. 
"logRatio" is currently the only option. See scoreDip function.}

\item{minReads}{Filter out bins with fewer than minReads reads. Only used
if there is a 'coverage' column in BSDT}
}
\value{
A MIRA score for each region set in GRList. See ?scoreDip.
}
\description{
Takes methylation data and sets of regions then aggregates
methylation for each region set and scores the resulting profile.
A wrapper for aggregateMethyl and scoreDip but it does not return
the summary methylation profiles, just the scores.
}
\examples{
data("exampleBSDT", package = "MIRA") 
data("exampleRegionSet", package = "MIRA") 
MIRAScore(BSDT = exampleBSDT, GRList = exampleRegionSet)

}
