% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{MIRAScore}
\alias{MIRAScore}
\title{Function to take DNA methylation and region data and return a MIRA score;
a wrapper for returnMIRABins and scoreDip}
\usage{
MIRAScore(BSDT, GRList, binNum = 11, scoringMethod = "logRatio",
  sampleName = NULL, sampleType = NULL)
}
\arguments{
\item{BSDT}{A single data table that has DNA methylation data including :.}

\item{GRList}{A GRangesList object containing region sets, each set corresponding to a regulatory element;
Each regionSet in the list should be named.}

\item{binNum}{How many bins each region should be split into for aggregation of the DNA methylation data.}

\item{scoringMethod}{Method to calculate MIRA score after binning; includes logratio, area; see scoreDip function.}

\item{sampleName}{}

\item{sampleType}{could be case/control, tissue type, etc.}
}
\description{
Function to take DNA methylation and region data and return a MIRA score;
a wrapper for returnMIRABins and scoreDip
}
