% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{parseBiseq}
\alias{parseBiseq}
\title{Takes a data.table from BSreadBiSeq and parses the strange x/y format
of methylation calls, splitting them into individual columns}
\usage{
parseBiseq(DT)
}
\arguments{
\item{DT}{data.table to parse}
}
\value{
data.table with separate methylated and unmethylated columns
}
\description{
Takes a data.table from BSreadBiSeq and parses the strange x/y format
of methylation calls, splitting them into individual columns
}
