% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMIRAProfiles}
\alias{plotMIRAProfiles}
\title{A function to plot the binned methylation of samples over a region.}
\usage{
plotMIRAProfiles(binnedRegDT, featID = unique(binnedRegDT[, featureID]),
  plotType = "line")
}
\arguments{
\item{binnedRegDT}{A datatable with specific column names containing:
bin numbers(binnedRegionDT column), 
aggregated methylation values (methylProp column), 
name of the region set (featureID column), 
case/control column (sampleType column), 
sample name (sampleName column).}

\item{featID}{Region set names in a single string or vector of strings.}

\item{plotType}{Line or jitter (ggplot2).}
}
\value{
A plot of class "gg"/ "ggplot" that shows MIRA signatures
}
\description{
A function to plot the binned methylation of samples over a region.
}
\examples{
data("exampleBins", package = "MIRA")
MIRAplot = plotMIRAProfiles(binnedRegDT = exampleBins)

}
