% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{plotMIRARegions}
\alias{plotMIRARegions}
\title{A function to plot the binned methylation of samples over a region.}
\usage{
plotMIRARegions(binnedRegDT, featID = unique(binnedRegDT[, featureID]),
  plotType = "line")
}
\arguments{
\item{binnedRegDT}{A datatable with specific column names containing the following:
bin numbers(binnedRegionDT), aggregated methylation values (methyl), name of the region set
(featureID), case/control column (sampleType), sample name (sampleName).}

\item{featID}{Region set names in a single string or vector of strings.}

\item{plotType}{Line or jitter (ggplot2).}
}
\description{
A function to plot the binned methylation of samples over a region.
}
