% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{plotMIRAScores}
\alias{plotMIRAScores}
\title{A function to plot MIRA scores and compare case/control.}
\usage{
plotMIRAScores(scoreDT, featID = unique(scoreDT[, featureID]))
}
\arguments{
\item{scoreDT}{A datatable with the following columns: score, featureID (names of regions),sampleType.}

\item{featID}{Region set name/names in a single string or vector of strings.}
}
\description{
A function to plot MIRA scores and compare case/control.
}

