% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMIRAScores}
\alias{plotMIRAScores}
\title{Plot MIRA scores and compare different conditions}
\usage{
plotMIRAScores(scoreDT, featID = unique(scoreDT[, featureID]),
  colBlindOption = FALSE)
}
\arguments{
\item{scoreDT}{A datatable with the following columns: 
score, featureID (names of regions), sampleType.}

\item{featID}{Region set name/names in a single string or vector of strings.}

\item{colBlindOption}{If TRUE, function will plot with a color blind
friendly palette which could be helpful when plotting multiple colors.}
}
\value{
a plot of class "gg"/"ggplot" that shows MIRA scores 
with geom_boxplot and geom_jitter.
}
\description{
Splits up samples by sample type. Displays each region set
in a different subplot.
If you only want to plot certain region sets,
subset with the `featID` parameter.
}
\examples{
data(ewingMyobigBinDT2) # bigBinDT object
exScores = bigBinDT2[, .(score=scoreDip(methylProp, 
                                       binCount = 21, 
                                       shoulderShift = 5)), 
                    by = .(featureID, sampleName)]
# adding annotation
sampleType = rep(c("Ewing", "Muscle-related"), each = 24)
exScores = cbind(exScores, sampleType)
exScorePlot = plotMIRAScores(exScores)         
}
