% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{returnMIRABins}
\alias{returnMIRABins}
\title{Function to aggregate methylation data into bins over all regions in each region set;}
\usage{
returnMIRABins(BSDT, GRList, binNum = 11, sampleNameInBSDT = TRUE,
  sampleType = NULL)
}
\arguments{
\item{BSDT}{A single data table that has DNA methylation data including :.}

\item{binNum}{How many bins each region should be split into for aggregation of the DNA methylation data}

\item{sampleType}{could be case/control, tissue type, etc.}

\item{GRDTList}{A GRangesList object containing region sets, each set corresponding to a regulatory element;
Each regionSet in the list should be named.}
}
\description{
Function to aggregate methylation data into bins over all regions in each region set;
}
