% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{scoreDip}
\alias{scoreDip}
\title{My dip scoring function - for MIRA scores;
That's Methylation-based Inference of Regulatory Activity}
\usage{
scoreDip(values, binCount, shoulderShift = 5, method = "logRatio")
}
\arguments{
\item{values}{A vector with proportion of methylation values for each bin. 
Between 0 and 1.}

\item{binCount}{Number of bins, also length of "values" vector.}

\item{shoulderShift}{The number of bins away from the center to use as the
shoulders. I have used 5 or 3.}

\item{method}{The scoring method. "logRatio" is the log of the ratio of
the average of outside values (shoulders) divided by the average of the middle values. 
A higher score with "logRatio" corresponds to a deeper dip.}
}
\value{
A MIRA score. The MIRA score quantifies the "dip" of the MIRA signature
which is an aggregation of methylation over all regions in a region set.
}
\description{
My dip scoring function - for MIRA scores;
That's Methylation-based Inference of Regulatory Activity
}
\examples{

#initial parameter
binNum=11

#loading sample annotation
load(system.file("data","annoDF.RData",package="MIRA"))
annoDF=as.data.table(annoDF)

#reading in region data and bisulfite data
someFeatures=get(load(system.file("data","Gm12878Nrf1_Subset.RData",package="MIRA")))
BSDT=get(load(system.file("data","GM06990_1_ExampleSet.RData",package = "MIRA")))

#converting to lists 
BSDTList=list(BSDT)
someFeatures=GRangesList(someFeatures)
names(someFeatures) <- "GM12878Nrf1"

#converting to proper format
BSDTList=addMethCol(BSDTList = BSDTList)


#doing MIRA analysis
bigBin=lapply(X = BSDTList,FUN = returnMIRABins,GRList=someFeatures,binNum=11,sampleNameInBSDT=TRUE)

bigBinDT=rbindlist(bigBin)#need to make sure that sample names have a column in here

#adding sampleType from annotation object
setkey(bigBinDT,sampleName)
setkey(annoDF,sampleName)
bigBinDT=merge(bigBinDT,annoDF, all.x=TRUE)

#scoring samples
sampleScores=bigBinDT[,.(score = scoreDip(methyl,binNum)),by=.(featureID,sampleName)]
setkey(sampleScores,sampleName)
setkey(annoDF,sampleName)
sampleScores=merge(sampleScores,annoDF,all.x=TRUE)

#visualizing MIRA signature
plotMIRARegions(binnedRegDT = bigBinDT)


}
