% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{scoreDip}
\alias{scoreDip}
\title{My dip scoring function - for MIRA scores;
That's Methylation-based Inference of Regulatory Activity
just calculates the ratio between the flank (shoulders) and midpoint}
\usage{
scoreDip(values, binCount, shoulderShift = 5, method = "logRatio")
}
\arguments{
\item{shoulderShift}{The number of bins away from the center to use as the
shoulders. I have used 5 or 3.}
}
\description{
My dip scoring function - for MIRA scores;
That's Methylation-based Inference of Regulatory Activity
just calculates the ratio between the flank (shoulders) and midpoint
}
