% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{scoreDip}
\alias{scoreDip}
\title{The dip scoring function for MIRA scores. This will take a vector
describing the methylation pattern and return a single score summarizing
that vector for how large the 'dip' in methylation is at the center of
the vector.}
\usage{
scoreDip(values, binCount, shoulderShift = "auto", method = "logRatio",
  usedStrand = FALSE)
}
\arguments{
\item{values}{A vector with proportion of methylation values for each bin. 
Between 0 and 1.}

\item{binCount}{Number of bins, also length of "values" vector.}

\item{shoulderShift}{Used to determine the number of bins away from the 
center to use as the shoulders. Default value "auto" optimizes the 
shoulderShift variable for each sample/region set combination to try find the 
outside edges of the dip. shoulderShift may be manually set as an integer
that will be used for all sample/region set combinations. "auto" does not 
currently work with region sets that include strand info.}

\item{method}{The scoring method. "logRatio" is the log of the ratio of outer
edges to the middle. This ratio is the average of outside values 
of the dip (shoulders) divided by the center value if 
it is lower than the two surrounding values or if it is not lower, an 
average of the three middle values. For an even binCount, the middle four
values would be averaged with the 1st and 4th being weighted by half (as
if there were 3 values). 
A higher score with "logRatio" corresponds to a deeper dip. "logRatio" is the
only scoring method currently but more methods may be added in the future.}

\item{usedStrand}{If strand information is included as part of an
input region set when aggregating methylation, 
the MIRA signature will probably not be 
symmetrical. In this case, the automatic 
shoulderShift sensing (done when shoulderShift="auto") needs to
be done for both sides of the dip instead of just one side so set
usedStrand=TRUE if strand was included for a region set.
usedStrand=TRUE only has an effect on the function when shoulderShift="auto".}
}
\value{
A MIRA score. The MIRA score quantifies the "dip" of 
the MIRA signature which is an aggregation of methylation 
over all regions in a region set.
}
\description{
The dip scoring function for MIRA scores. This will take a vector
describing the methylation pattern and return a single score summarizing
that vector for how large the 'dip' in methylation is at the center of
the vector.
}
\examples{
data("exampleBins")
binCount = 11 # bin number for exampleBins 
exampleBins[, .(score = scoreDip(methylProp, binCount)), 
              by = .(featureID, sampleName)]
}
