% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRA.R
\name{scoreDip}
\alias{scoreDip}
\title{My dip scoring function - for MIRA scores;
That's Methylation-based Inference of Regulatory Activity}
\usage{
scoreDip(values, binCount, shoulderShift = 5, method = "logRatio")
}
\arguments{
\item{values}{A vector with proportion of methylation values for each bin. 
Between 0 and 1.}

\item{binCount}{Number of bins, also length of "values" vector.}

\item{shoulderShift}{The number of bins away from the center to use as the
shoulders. I have used 5 or 3.}

\item{method}{The scoring method. "logRatio" is the log of the ratio of
the average of outside values (shoulders) divided by the average of the middle values. 
A higher score with "logRatio" corresponds to a deeper dip. "logRatio" is the
only scoring method currently but more methods may be added in the future.}
}
\value{
A MIRA score. The MIRA score quantifies the "dip" of the MIRA signature
which is an aggregation of methylation over all regions in a region set.
}
\description{
My dip scoring function - for MIRA scores;
That's Methylation-based Inference of Regulatory Activity
}
\examples{
data("exampleBins")
binCount=11 #bin number for exampleBins 
exampleBins[,scoreDip(methyl,binCount),by=.(featureID,sampleName)]
}
